Create DialogueManager
When the player is close enough to an NPC, by pressing E they can enter dialogue with them. By using the W and S buttons, they can scroll between the dialogue options and press 'return' to get the intended response.

I have made it so that when the player enters dialogue, their speeds are set to 0. This stops them from being able to move during dialogue which allows the movement buttons to be used as the dialogue option buttons which I think is just more comfortable for the player. When dialogue is exited, they get their speed back, allowing them to move again.